Disable workaround for apply-special-only-once if RULE_APPLY_ALL_EFFECTS_AFTER_PRODUCTION
The fuel special wants once to set all ships in the system to maximum.
Without RULE_APPLY_ALL_EFFECTS_AFTER_PRODUCTION the special needs to exist
at the after-production because the fuel change in before-production gets
reset. To be able to remove the apply-once special, as a workaround another
special was added to the ships, which triggered the fuel change in after-production.
That ship specials removed itself in the before-production of the following turn
and the corresponding meter change was reset after before-production.