buildings: colonies: fix exticts species colonies
The format conversion for colony building generation altered the logic
for colonization. This makes it impossible to colonize planets with
extinct species like "The Banforo".

For the colonization to be possible it should EITHER be required the
planet to be colonized has a supply-connection to a planet which:

1. Has already colony of the species
	1a) Which has sufficient population
	2b) Is happy enough

OR

2. Has the extict species special + required tech and a Xenoresurrection
   lab.

The conversion altered this condition to require both, which is not
possible.

Fix the extinct species resurrection by requiring only one of the above.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
Fixes: 58efaab1f ("[FOCS] Convert colonies buildings (#5185)")