Update NEUTRONIUM_FORGE.focs.py (#5302)
In the late game, you can queue an entire "production pipeline" in one go - except the Neutronium Forge, which has to wait for the first step to be completed. This "fixes" that slight perceived inconsistency. 

Not I didn't try to allow the queueing once the extractor is queued, that wouldn't be so trivially simple...